урок 6 настроили работу IntentService+BroadcastReceiver

+++ app/src/main/java/com/gb/kotlin_1728_2_1/model/WeatherDTO.kt
@@ -103,7 +103,7 @@ data class Part (
     val humidity: Long,
 
     @SerializedName("prec_mm")
-    val precMm: Long,
+    val precMm: Double,
 
     @SerializedName("prec_prob")
     val precProb: Long,
